requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-suit==0.2.26
djangorestframework==3.8.2
drf-nested-routers==0.90.2
idna==2.6
pytz==2018.4
requests==2.18.4
urllib3==1.22